fix(dbns): keep range() off the iostream layer and the conversion trace constexpr-safe
CodeRabbit review on #1449. Both confirmed by building the failing case first.

range() lives in manipulators.hpp and streamed dbns values into a stringstream,
so instantiating it needed operator<<(ostream, dbns) -- which this PR moves to
iostream.hpp. A translation unit including only core.hpp + manipulators.hpp
therefore failed to compile the moment range() was used, which makes the layer
boundary this PR claims simply untrue.

The old in-class operator<< was `ostr << double(r)`, so streaming double(b)
directly is the same text by construction, and the differential confirms it.

MY LAYER GATE MISSED THIS AND HAS BEEN FIXED. It compiled manipulators.hpp but
never instantiated anything in it, and a template body is not checked until
instantiated -- the same class of blind spot as gating on the umbrella building,
one level down. The gate TU now calls type_tag, type_field, range, to_binary
(both nibble settings), to_hex, pretty_print, info_print and color_print.

The DBNS_TRACE_CONVERSION sites needed !std::is_constant_evaluated(). This PR
turns three `if constexpr (bDebug)` blocks, where bDebug was permanently false,
into real preprocessor-guarded code. With tracing switched on, an unguarded
std::cout inside convert_ieee754 makes the function non-constant-evaluable, so
`constexpr dbns x(1.5)` stops compiling. The guard is the idiom the statistics
counters in the same function already use. Runtime tracing is unaffected:
is_constant_evaluated() is false there.

Not fixed, filed as #1450: operator>> extracts into an uninitialised double and
assigns it unconditionally. Real, but confined to an already-failed stream (on a
good stream num_get stores 0 per C++11), pre-existing verbatim from the in-class
friend on main, and the fix changes the common failure path -- so it needs a
BEHAVIOUR-CHANGE trailer and a decision about what extraction failure should
leave behind, given Universal's types are deliberately trivially constructible.

Re-verified after both fixes: 29,992,863 bytes of differential output still
byte-identical to main on gcc and clang and between them, the trace comparison
still 1,273,937 bytes identical, a constexpr dbns now compiles with tracing on,
and the dbns targets still pass on both compilers.

+++ include/sw/universal/number/dbns/manipulators.hpp
@@ -54,7 +54,11 @@ namespace sw { namespace universal {
 	inline std::string range(const DbnsType & = {}) {
 		std::stringstream s;
 		DbnsType b(SpecificValue::maxneg), c(SpecificValue::minneg), d(SpecificValue::minpos), e(SpecificValue::maxpos);
-		s << "[" << b << " ... " << c << ", 0, " << d << " ... " << e << "]\n";
+		// stream the double conversions rather than the dbns values themselves: this
+		// layer must not depend on iostream.hpp (#1334), and operator<<(ostream, dbns)
+		// lives there. It is `ostr << double(r)`, so this is the same text -- and the
+		// same text is what the differential in the PR checks.
+		s << "[" << double(b) << " ... " << double(c) << ", 0, " << double(d) << " ... " << double(e) << "]\n";
 		return s.str();
 	}
